% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls.R
\name{cls}
\alias{cls}
\title{Compute linear regression with constraints}
\usage{
cls(y, Z, L, c, intercept = T)
}
\arguments{
\item{y}{tba}

\item{Z}{tba}

\item{L}{tba}

\item{c}{tba}

\item{intercept}{tba}
}
\value{
A list
}
\description{
Compute the least square solution from linear regression with constraint 
\eqn{L*beta=c}.
}
