% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_L.R
\name{comp_L}
\alias{comp_L}
\title{Constraint matrix computation}
\usage{
comp_L(p_vec)
}
\arguments{
\item{p_vec}{A vector of length \eqn{q} whose elements are the number of 
components of each composition.}
}
\value{
The matrix \eqn{L}.
}
\description{
Compute the matrix representing the set of linear constraints.
}
