% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_comp.R
\name{error_comp}
\alias{error_comp}
\title{Error computation}
\usage{
error_comp(beta, betac, Y, Z, Phi, t, Zc = NULL)
}
\arguments{
\item{beta}{The estimated \eqn{beta} (for a single \eqn{lambda})}

\item{betac}{The estimated \eqn{betac} (for a single \eqn{lambda}).}

\item{Y}{A \eqn{n_t*n} array, containing \eqn{n} samples of the functional 
response at \eqn{n_t} points.}

\item{Z}{A \eqn{n_t*p*n} array, containing \eqn{n} samples of \eqn{p} functional
predictors observed at \eqn{n_t} points.}

\item{Phi}{The matrix \eqn{Phi} from \eqn{mat_comp}.}

\item{t}{A vector containing the values at which functions are observed.}

\item{Zc}{optional. An array of \eqn{p_c} control variables, with same structure of Z. The
first predictor must be a vector of ones. The default value is an array of ones,
to estimate the intercept.}
}
\value{
A list containing the MSE and the APMSE.
}
\description{
Compute the mean square error (MSE) and the average prediction MSE (APMSE).
}
