% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcls.R
\name{fcls}
\alias{fcls}
\title{Compute functional concurrent regression with constraints}
\usage{
fcls(matrices, Ltilda, c)
}
\arguments{
\item{matrices}{tba.}

\item{Ltilda}{tba.}

\item{c}{tba.}
}
\value{
A list.
}
\description{
Compute the least square solution from functional concurrent regression with constraint 
\eqn{L*beta=c}.
}
\details{
The Lagrangian is \eqn{L(beta, u) = -t(y)*x*beta+t(beta)*t(x)*x*beta+t(u)*(Ltilda*beta-c)}.
}
