\name{fcrc-package}
\alias{fcrc-package}
\alias{fcrc}
\docType{package}
\title{
\packageTitle{fcrc}
}
\description{
\packageDescription{fcrc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fcrc}
\packageIndices{fcrc}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{fcrc}

Maintainer: \packageMaintainer{fcrc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
