% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcrc.R
\docType{package}
\name{fcrc}
\alias{fcrc}
\title{fcrc: A package for estimating a functional concurrent regression model with 
compositional covariates.}
\description{
The package contains the functions for estimating the model proposed in the 
paper \href{https://arxiv.org/abs/2301.06333}{“Functional Concurrent Regression with Compositional Covariates 
and its application to the time-varying effect of causes of death on 
human longevity”} by E. G. Depaoli, M. Stefanucci and S. Mazzuco.
}
