% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_comp.R
\name{mat_comp}
\alias{mat_comp}
\title{Matrix computation for model estimation}
\usage{
mat_comp(Z, Y, t, k, Zc = NULL)
}
\arguments{
\item{Z}{A \eqn{n_t*p*n} array, containing \eqn{n} samples of \eqn{p} functional
predictors observed at \eqn{n_t} points.}

\item{Y}{A \eqn{n_t*n} array, containing \eqn{n} samples of the functional 
response at \eqn{n_t} points.}

\item{t}{A vector containing the values at which functions are observed.}

\item{k}{The number of basis function.}

\item{Zc}{An optional array of \eqn{p_c} control variables, with same structure of Z. The
first predictor must be a vector of ones.}
}
\value{
A list containing the matrices \eqn{J}, \eqn{K}, \eqn{M}, \eqn{P} and 
        other quantities used for model estimation.
}
\description{
Compute the matrices required for estimating the functional model, using B-splines
of order 4 as basis functions.
}
