% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_comp_cgl.R
\name{path_comp_cgl}
\alias{path_comp_cgl}
\title{Compute regularization path}
\usage{
path_comp_cgl(matrices)
}
\arguments{
\item{matrices}{A list from \link{mat_comp}.}
}
\value{
A vector with the regularization path.
}
\description{
Compute the regularization path to be used by \link{alm_cgl_path}.
}
