% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cv.R
\name{plot_cv}
\alias{plot_cv}
\title{Plot the results of the n-fold cross-validation}
\usage{
plot_cv(results)
}
\arguments{
\item{results}{A list from \url{cv_cgl}.}
}
\description{
Plot the results of the n-fold cross-validation
}
