% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapz.R
\name{trapz}
\alias{trapz}
\title{Trapezoidal Integration}
\usage{
trapz(x, y)
}
\arguments{
\item{x}{\eqn{x}-coordinates of points.}

\item{y}{\eqn{y}-coordinates of points.}
}
\value{
Approximated integral of the function.
}
\description{
Compute the area of a function with values \eqn{y} at points \eqn{x}, using the
trapezoidal rule.
}
