% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alm_cgl_path.R
\name{alm_cgl_path}
\alias{alm_cgl_path}
\title{Compute constrained group Lasso}
\usage{
alm_cgl_path(
  matrices,
  L,
  lambda_path = NULL,
  beta_start = NULL,
  u_start = NULL,
  rho = 1,
  max_rho = 1e+06,
  gamma = 1,
  varying_gamma = T,
  eps = 1e-05,
  max_iter_int = 1000,
  max_iter = 100,
  abs_tol_int = 1e-06,
  rel_tol_int = 1e-04,
  zero_tol = 1e-04
)
}
\description{
tba
}
\details{
tba
}
