% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_cgl.R
\name{cv_cgl}
\alias{cv_cgl}
\title{Compute n-fold cross validation}
\usage{
cv_cgl(
  Y,
  Z,
  L,
  t,
  nfold = 10,
  k_range = 4:10,
  Zc = NULL,
  ncores = NULL,
  eps = 1e-05,
  max_iter_int = 1000,
  max_iter = 100,
  abs_tol_int = 1e-06,
  rel_tol_int = 1e-04,
  zero_tol = 1e-04
)
}
\description{
Compute the regularization path for constrained group Lasso.
}
\details{
tba
}
