% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_comp.R
\name{error_comp}
\alias{error_comp}
\title{Error computation}
\usage{
error_comp(beta, betac, Y, Z, Phi, t, Zc = NULL)
}
\arguments{
\item{beta}{to be added}

\item{betac}{to be added}

\item{Y}{to be added}

\item{Z}{to be added}

\item{Phi}{to de added}

\item{t}{to be added}

\item{Zc}{to be added}
}
\value{
A list containing the MSE and the APMSE.
}
\description{
Compute the mean square error (MSE) and the average prediction MSE (APMSE).
}
