% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_comp_cgl.R
\name{path_comp_cgl}
\alias{path_comp_cgl}
\title{Compute regularization path}
\usage{
path_comp_cgl(matrices)
}
\arguments{
\item{matrices}{tba.}
}
\value{
A vector with the regularization path.
}
\description{
Compute the regularization path for constrained group Lasso.
}
